requirements.txt
openai>=1.0.0
Pillow>=10.0.0
pywin32>=306
pyautogui>=0.9.54
pyperclip>=1.8.2
customtkinter>=5.2.0
comtypes>=1.4.0
